stage-1/pyproject.toml
[project]
name = "stage-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]
